Clean and refactor.
