+++ src/main/java/com/affecto/jsonstat/dto/IndividualDimensionBlock.java
@@ -1,0 +1,12 @@
+package com.affecto.jsonstat.dto;
+
+import lombok.Data;
+
+@Data
+public class IndividualDimensionBlock {
+
+    public String label;
+
+    public DimensionCategoryBlock category;
+
+}
