+++ src/main/java/com/affecto/jsonstat/dto/IndividualDimensionUnitBlock.java
@@ -1,0 +1,16 @@
+package com.affecto.jsonstat.dto;
+
+import lombok.Data;
+
+@Data
+public class IndividualDimensionUnitBlock {
+
+    public String type;
+
+    public String base;
+
+    public String symbol;
+
+    public Integer multiplier;
+
+}
