+++ src/test/java/com/affecto/jsonstat/DeserializeTests.java
@@ -1,6 +1,8 @@
 package com.affecto.jsonstat;
 
+import com.affecto.jsonstat.dto.DimensionGroupBlock;
 import com.affecto.jsonstat.dto.JsonStat;
+import com.affecto.jsonstat.util.DimensionGroupBlockDeserializer;
 import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
@@ -16,7 +18,7 @@ public class DeserializeTests {
         final ObjectMapper om = new ObjectMapper();
         om.registerModule(new JavaTimeModule());
         SimpleModule testModule = new SimpleModule("MyModule", new Version(1, 0, 0, null))
-                     .addDeserializer(JsonStat.DimensionGroupBlock.class, new JsonStat.DimensionGroupBlockDeserializer());
+                     .addDeserializer(DimensionGroupBlock.class, new DimensionGroupBlockDeserializer());
         om.registerModule(testModule);
         try (final InputStream is = DeserializeTests.class.getClassLoader().getResourceAsStream("us-labor-ds.json")) {
             final JsonStat uld = om.readValue(is, JsonStat.class);
